🧢 Add size, padding, label, position. Refactor toogle;
